@@ -1,14 +1,14 @@
-import { NgxDynamicTemplateDemoPage } from './app.po';
+import { AppPage } from './app.po';
 
-describe('ngx-dynamic-component-demo App', () => {
-  let page: NgxDynamicTemplateDemoPage;
+describe('ngx-dynamic-template-demo App', () => {
+  let page: AppPage;
 
   beforeEach(() => {
-    page = new NgxDynamicTemplateDemoPage();
+    page = new AppPage();
   });
 
-  it('should display message saying app works', () => {
+  it('should display welcome message', () => {
     page.navigateTo();
-    expect(page.getParagraphText()).toEqual('app works!');
+    expect(page.getParagraphText()).toEqual('Welcome to app!');
   });
 });
